requirements.txt
absl-py==2.1.0
APScheduler==3.11.2
astunparse==1.6.3
beautifulsoup4==4.14.3
blinker==1.9.0
cachelib==0.13.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.3.0
colorama==0.4.6
comtypes==1.4.12
Deprecated==1.3.1
docopt==0.6.2
ecdsa==0.19.1
Flask==3.1.2
Flask-Caching==2.3.1
flask-cors==6.0.2
Flask-Limiter==4.0.0
flask-marshmallow==1.3.0
Flask-SQLAlchemy==3.1.1
flask-swagger==0.2.14
flask-swagger-ui==5.21.0
flatbuffers==25.1.24
gast==0.6.0
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.70.0
gunicorn==23.0.0
h5py==3.12.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Js2Py==0.74
keras==3.8.0
libclang==18.1.1
limits==5.6.0
Markdown==3.7
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.1.0
marshmallow-sqlalchemy==1.4.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==2.0.2
opt_einsum==3.4.0
optree==0.14.0
ordered-set==4.1.0
packaging==25.0
pipwin==0.5.2
protobuf==5.29.3
psycopg2==2.9.11
pyasn1==0.6.1
pycoingecko==3.2.0
pycryptodome==3.21.0
pygame==2.6.1
Pygments==2.19.2
pyjsparser==2.7.1
PyJWT==2.10.1
PyPDF2==3.0.1
pypiwin32==223
PyPrind==2.11.3
pySmartDL==1.3.4
python-dotenv==1.2.1
python-jose==3.5.0
pyttsx3==2.99
pywin32==311
PyYAML==6.0.3
requests==2.32.3
rich==14.2.0
rsa==4.9.1
six==1.17.0
soupsieve==2.8
SQLAlchemy==2.0.44
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
tk==0.1.0
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==2.0.1
